Move status condition enums into api package
